backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pgvector
python-dotenv
requests
beautifulsoup4
sentence-transformers
konlpy
passlib[bcrypt]
google-generativeai
